controllers and routes implemented
